// src/lib/api/services/QuotesService/QuotesService.ts
export interface QuotesQueryParams {
  userId: number;
  companyId: number;
  offset?: number;
  limit?: number;
}

// Quotes Request Body Interface
export interface QuotesRequestBody {
  filter_index: number;
  filter_name: string;
  endCreatedDate: string;
  endDate: string;
  endValue: string;
  endTaxableAmount: string;
  endGrandTotal: string;
  identifier: string;
  limit: number;
  offset: number;
  name: string;
  pageNumber: number;
  startDate: string;
  startCreatedDate: string;
  startValue: string;
  startTaxableAmount: string;
  startGrandTotal: string;
  status: string[];
  selectedColumns: string[];
  columnWidth: number[];
  columnPosition: string;
  userDisplayName: string;
  userStatus: string[];
  accountId: number[];
  branchId: number[];
}
// ...
export interface QuoteItem {
  SPRRequested: boolean;
  approvalAwaiting: unknown | null;
  approvalGroupId: ApprovalGroup | null;
  approvalInitiated: boolean;
  buyerCompanyBranchName: string;
  buyerCompanyName: string;
  calculatedTotal: number;
  createdDate: string;
  curencySymbol: CurrencySymbol;
  customerRequiredDate: string | null;
  erpCode: string | null;
  grandTotal: number;
  itemCount: number;
  lastUpdatedDate: string;
  leadIdentifier: string | null;
  orderIdentifier: string[] | null;
  quotationIdentifier: string;
  quoteName: string;
  quoteUsers: QuoteUser[] | null;
  sellerCompanyBranchName: string;
  sellerCompanyName: string;
  subTotal: number;
  taxableAmount: number;
  updatedBuyerStatus: string;
  updatedSellerStatus: string;
  vendorId: number | null;
}

// Quotes Response Data Interface
export interface QuotesResponseData {
  quotesResponse: QuoteItem[];
  totalQuoteCount: number;
}

// Quotes Response Interface
export interface QuotesApiResponse {
  data: QuotesResponseData;
  message: string | null;
  status: string;
}